Jeickssonn

+++ app/Http/Controllers/API/CategoriaApiController.php
@@ -16,8 +16,10 @@ class CategoriaApiController extends Controller
      */
     public function index()
     {
-        $categorias = Categoria::all();
-        return response()->json(['data' => $categorias], 200);
+        //$categoria = Categoria::with('producto')->get();
+        $categoria = Categoria::all();
+        return response()->json(['Categoria' => $categoria]);
+
     }
 
 
@@ -30,12 +32,15 @@ class CategoriaApiController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'nombre' => 'required|string',
-            'descripcion' => 'required',
+            'nombre' => 'required|string'
         ]);
 
-        $categoria = Categoria::create($request->all());
-        return response()->json(['data' => $categoria], 201);
+        $categoria = Categoria::create([
+            'nombre' => $request->nombre
+
+        ]);
+
+        return response()->json(['Categoria' => $categoria], 201);
     }
 
 
@@ -45,9 +50,11 @@ class CategoriaApiController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(Categoria $categoria)
+    public function show($id)
     {
-        return response()->json(['data' => $categoria], 200);
+        $categoria = Categoria::find($id);
+        return response()->json($categoria, 200);
+
     }
 
 
@@ -58,15 +65,24 @@ class CategoriaApiController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Categoria $categoria)
+    public function update(Request $request, $id)
     {
-        $request->validate([
-            'nombre' => 'required|string',
-            'descripcion' => 'required',
+        $this->validate($request, [
+            'nombre' => 'required|string'
+        ]); 
+
+        $categoria = Categoria::find($id);
+
+        if(!$categoria){
+            return response()->json(['error' => 'Categoria no encontrada'], 404);
+        }
+
+        $categoria->update([
+            'nombre' => $request->nombre
         ]);
 
-        $categoria->update($request->all());
-        return response()->json(['data' => $categoria], 200);
+        return response()->json(['message' => 'Categoria actualizada', 'Categoria'=>$categoria]);
+
     }
 
 
@@ -76,10 +92,19 @@ class CategoriaApiController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Categoria $categoria)
+    public function destroy($id)
     {
+
+        $categoria = Categoria::find($id);
+
+        if(!$categoria){
+            return response()->json(['error' => 'Categoria no encontrada'], 404);
+        }
+        // eliminamos el producto
         $categoria->delete();
-        return response(null, 204);
+
+        return response()->json(['message'=>'Categoria eliminada'], 200);
+
     }
 
 }
